Improved default scene: skysphere + post-processing + better grid (#511)
* Added skysphere and updated default scene
* Fixed missing hdr extension from asset importer
* Fixed HDR texture were 16bits while 32bits was expected
* Cleaned up grid shader
* Enabled default skysphere in asset view
* Added options to create skysphere shader and material
* Reworked `CreateNewMaterial` to not use hard coded XML, and updated skysphere material creation to not use backface culling
* Changed default grid color to 100% white, and adjusted grid fading
* Remove grid render pass stencil clear (as it doesnt use it anymore) and added it to DebugSceneRenderer for outline
* Update grid line width and fading

+++ Sources/Overload/OvRendering/src/OvRendering/Resources/Loaders/TextureLoader.cpp
@@ -63,7 +63,7 @@ namespace
 			.height = p_height,
 			.minFilter = p_minFilter,
 			.magFilter = p_magFilter,
-			.internalFormat = p_hdr ? EInternalFormat::RGBA16F : EInternalFormat::RGBA8,
+			.internalFormat = p_hdr ? EInternalFormat::RGBA32F : EInternalFormat::RGBA8,
 			.useMipMaps = p_generateMipmap
 		});
